add scripts to deploy router and pool
