+++ scripts/zilswap-v2/deployPool.js
@@ -13,30 +13,14 @@ getAmpBps = (isAmpPool) => {
 
 ;
 (async () => {
-  const privateKey = getPrivateKey()
-  const ownerAddress = getAddressFromPrivateKey(privateKey);
+  const privateKey = getPrivateKey();
   const wZilAddress = process.env.WZIL_CONTRACT_HASH;
   const wZilContract = zilliqa.contracts.at(wZilAddress);
   const hunyAddress = process.env.HUNY_CONTRACT_HASH;
   const hunyContract = zilliqa.contracts.at(hunyAddress);
-  const codehash = getContractCodeHash("./src/zilswap-v2/ZilSwapPool.scilla");
 
-  const routerContract = (await deployZilswapV2Router(privateKey, { governor: null, codehash, wZil: wZilAddress.toLowerCase()}))[0];
-  const routerAddress = routerContract.address.toLowerCase()
-
-  const txSetFeeConfig = await callContract(
-    privateKey, routerContract,
-    'SetFeeConfiguration',
-    [
-      param('config', 'Pair ByStr20 Uint128', {
-        "constructor": "Pair",
-        "argtypes": ["ByStr20", "Uint128"],
-        "arguments": [`${ownerAddress}`, "1000"] // 10%
-      })
-    ],
-    0, false, false
-  );
-  console.log(`setting fee configuration on router: ${routerAddress}`, txSetFeeConfig.id);
+  const routerAddress = process.env.ROUTER_CONTRACT_HASH;
+  const routerContract = zilliqa.contracts.at(routerAddress);
 
   const poolContract = (await deployZilswapV2Pool(privateKey, { factory: routerContract, token0: wZilContract, token1: hunyContract, init_amp_bps: getAmpBps(false)}))[0];
   const poolAddress = poolContract.address.toLowerCase();
@@ -54,6 +38,5 @@ getAmpBps = (isAmpPool) => {
   console.log(`\n\n======================`)
   console.log(`\n  Contracts`)
   console.log(`\n======================`)
-  console.log(`\nRouter      `, routerAddress);
   console.log(`\nPool        `, poolAddress);
 })();
